backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=1.6.0
pydantic-settings>=2.0.0
httpx>=0.25.0
slowapi>=0.1.9
websockets>=12.0
pyyaml>=6.0.0
openai>=1.0.0
anthropic>=0.40.0
openai-agents @ git+https://github.com/openai/openai-agents-python
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
cryptography>=42.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
